@@ -75,8 +75,8 @@ compileTestJava {
 }
 
 java {
-  sourceCompatibility = JavaVersion.VERSION_11
-  targetCompatibility = JavaVersion.VERSION_11
+  sourceCompatibility = JavaVersion.VERSION_21
+  targetCompatibility = JavaVersion.VERSION_21
 }
 
 tasks.withType(JavaCompile) {
@@ -86,5 +86,6 @@ tasks.withType(JavaCompile) {
 }
 
 test {
+  jvmArgs '--add-opens=java.base/java.lang=ALL-UNNAMED' // for mockito reflection
   systemProperty "java.util.logging.manager", "org.jboss.logmanager.LogManager"
 }
