@@ -1,6 +1,6 @@
 package net.explorviz.adapter.service.converter;
 
-import javax.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.context.ApplicationScoped;
 import net.explorviz.avro.Span;
 
 /**
