@@ -3,8 +3,9 @@ package net.explorviz.adapter.injection;
 import io.confluent.kafka.schemaregistry.client.MockSchemaRegistryClient;
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
 import io.quarkus.arc.profile.IfBuildProfile;
-import javax.enterprise.context.Dependent;
-import javax.enterprise.inject.Produces;
+import jakarta.enterprise.context.Dependent;
+import jakarta.enterprise.inject.Produces;
+
 
 @Dependent
 public class MockSchemaRegistryClientProducer {
